Common Data Set Extracts

Extracted portions of school Common Data Set reports.

Usage

cds_data

Format

A data.frame with a report in each row, and fields in each column. A few columns contain metadata:

school

Name of the school.

source

URL of the original PDF file.

md5

MD5 hash of the original PDF file.

retieved

Date downloaded.

processed

Date processed.

year

Year of the report.

The rest are named first by section and item number (e.g., B1), then by a combination of levels such as the part, column name, and row name from the original table.

Examples

cds_data[1:10, c(
  "school", "retrieved", "B1_men_undergrad_new_full", "B1_women_undergrad_new_full"
)]
#>                 school                retrieved B1_men_undergrad_new_full
#> 1   Belmont University Wed Oct  9 07:42:44 2024                        NA
#> 2   Belmont University Wed Oct  9 07:42:44 2024                        NA
#> 3   Belmont University Wed Oct  9 07:42:46 2024                        NA
#> 4   Belmont University Wed Oct  9 07:42:47 2024                        NA
#> 5   Belmont University Wed Oct  9 07:42:47 2024                       519
#> 6   Belmont University Wed Oct  9 07:42:47 2024                        NA
#> 7   Belmont University Wed Oct  9 07:42:48 2024                       491
#> 8   Belmont University Wed Oct  9 07:42:48 2024                        NA
#> 9   Belmont University Wed Oct  9 07:42:48 2024                       502
#> 10 Brandeis University Wed Oct  9 07:42:49 2024                        NA
#>    B1_women_undergrad_new_full
#> 1                            1
#> 2                          607
#> 3                           NA
#> 4                          514
#> 5                           NA
#> 6                           NA
#> 7                           NA
#> 8                          480
#> 9                          752
#> 10                          NA