Process a CDS Report

Extract data from a CDS PDF report.

Usage

cds_process_report(school, file, ocr = FALSE)

Arguments

school

Name of the school, to be checked for in the document.

file

Path to the PDF report.

ocr

Logical; if TRUE, will use Optical Character Recognition to extract text from the file.

Value

A list containing the locations of items identified in the report and the values extracted from the report.

Examples

file <- "../../reports/text/Yale University/cds2008_2009_1.txt"
if (file.exists(file)) {
  values <- cds_process_report("Yale University", file)
  values[1:10]
}
#> $locations
#> $locations$A
#> $locations$A$start
#> [1] 4
#> 
#> $locations$A$end
#> [1] 46
#> 
#> $locations$A$items
#> $locations$A$items$A1
#> $locations$A$items$A1$start
#> [1] 4
#> 
#> $locations$A$items$A1$end
#> [1] 22
#> 
#> 
#> $locations$A$items$A2
#> $locations$A$items$A2$start
#> [1] 23
#> 
#> $locations$A$items$A2$end
#> [1] 26
#> 
#> 
#> $locations$A$items$A3
#> $locations$A$items$A3$start
#> [1] 27
#> 
#> $locations$A$items$A3$end
#> [1] 30
#> 
#> 
#> $locations$A$items$A4
#> $locations$A$items$A4$start
#> [1] 31
#> 
#> $locations$A$items$A4$end
#> [1] 35
#> 
#> 
#> $locations$A$items$A5
#> $locations$A$items$A5$start
#> [1] 36
#> 
#> $locations$A$items$A5$end
#> [1] 46
#> 
#> 
#> 
#> 
#> $locations$B
#> $locations$B$start
#> [1] 47
#> 
#> $locations$B$end
#> [1] 161
#> 
#> $locations$B$items
#> $locations$B$items$B1
#> $locations$B$items$B1$start
#> [1] 47
#> 
#> $locations$B$items$B1$end
#> [1] 77
#> 
#> 
#> $locations$B$items$B2
#> $locations$B$items$B2$start
#> [1] 78
#> 
#> $locations$B$items$B2$end
#> [1] 94
#> 
#> 
#> $locations$B$items$B3
#> $locations$B$items$B3$start
#> [1] 95
#> 
#> $locations$B$items$B3$end
#> [1] 120
#> 
#> 
#> $locations$B$items$B4
#> $locations$B$items$B4$start
#> [1] 121
#> 
#> $locations$B$items$B4$end
#> [1] 123
#> 
#> 
#> $locations$B$items$B5
#> $locations$B$items$B5$start
#> [1] 124
#> 
#> $locations$B$items$B5$end
#> [1] 128
#> 
#> 
#> $locations$B$items$B6
#> $locations$B$items$B6$start
#> [1] 129
#> 
#> $locations$B$items$B6$end
#> [1] 131
#> 
#> 
#> $locations$B$items$B7
#> $locations$B$items$B7$start
#> [1] 132
#> 
#> $locations$B$items$B7$end
#> [1] 134
#> 
#> 
#> $locations$B$items$B8
#> $locations$B$items$B8$start
#> [1] 135
#> 
#> $locations$B$items$B8$end
#> [1] 137
#> 
#> 
#> $locations$B$items$B9
#> $locations$B$items$B9$start
#> [1] 138
#> 
#> $locations$B$items$B9$end
#> [1] 141
#> 
#> 
#> $locations$B$items$B10
#> $locations$B$items$B10$start
#> [1] 142
#> 
#> $locations$B$items$B10$end
#> [1] 143
#> 
#> 
#> $locations$B$items$B11
#> $locations$B$items$B11$start
#> [1] 144
#> 
#> $locations$B$items$B11$end
#> [1] 153
#> 
#> 
#> $locations$B$items$B22
#> $locations$B$items$B22$start
#> [1] 154
#> 
#> $locations$B$items$B22$end
#> [1] 161
#> 
#> 
#> 
#> 
#> $locations$C
#> $locations$C$start
#> [1] 162
#> 
#> $locations$C$end
#> [1] 441
#> 
#> $locations$C$items
#> $locations$C$items$C1
#> $locations$C$items$C1$start
#> [1] 162
#> 
#> $locations$C$items$C1$end
#> [1] 175
#> 
#> 
#> $locations$C$items$C2
#> $locations$C$items$C2$start
#> [1] 176
#> 
#> $locations$C$items$C2$end
#> [1] 185
#> 
#> 
#> $locations$C$items$C3
#> $locations$C$items$C3$start
#> [1] 186
#> 
#> $locations$C$items$C3$end
#> [1] 190
#> 
#> 
#> $locations$C$items$C4
#> $locations$C$items$C4$start
#> [1] 191
#> 
#> $locations$C$items$C4$end
#> [1] 197
#> 
#> 
#> $locations$C$items$C5
#> $locations$C$items$C5$start
#> [1] 198
#> 
#> $locations$C$items$C5$end
#> [1] 212
#> 
#> 
#> $locations$C$items$C6
#> $locations$C$items$C6$start
#> [1] 213
#> 
#> $locations$C$items$C6$end
#> [1] 220
#> 
#> 
#> $locations$C$items$C7
#> $locations$C$items$C7$start
#> [1] 221
#> 
#> $locations$C$items$C7$end
#> [1] 248
#> 
#> 
#> $locations$C$items$C8
#> $locations$C$items$C8$start
#> [1] 249
#> 
#> $locations$C$items$C8$end
#> [1] 305
#> 
#> 
#> $locations$C$items$C9
#> $locations$C$items$C9$start
#> [1] 306
#> 
#> $locations$C$items$C9$end
#> [1] 343
#> 
#> 
#> $locations$C$items$C10
#> $locations$C$items$C10$start
#> [1] 344
#> 
#> $locations$C$items$C10$end
#> [1] 356
#> 
#> 
#> $locations$C$items$C11
#> $locations$C$items$C11$start
#> [1] 357
#> 
#> $locations$C$items$C11$end
#> [1] 368
#> 
#> 
#> $locations$C$items$C12
#> $locations$C$items$C12$start
#> [1] 369
#> 
#> $locations$C$items$C12$end
#> [1] 372
#> 
#> 
#> $locations$C$items$C13
#> $locations$C$items$C13$start
#> [1] 373
#> 
#> $locations$C$items$C13$end
#> [1] 384
#> 
#> 
#> $locations$C$items$C14
#> $locations$C$items$C14$start
#> [1] 385
#> 
#> $locations$C$items$C14$end
#> [1] 388
#> 
#> 
#> $locations$C$items$C15
#> $locations$C$items$C15$start
#> [1] 389
#> 
#> $locations$C$items$C15$end
#> [1] 389
#> 
#> 
#> $locations$C$items$C16
#> $locations$C$items$C16$start
#> [1] 390
#> 
#> $locations$C$items$C16$end
#> [1] 393
#> 
#> 
#> $locations$C$items$C17
#> $locations$C$items$C17$start
#> [1] 394
#> 
#> $locations$C$items$C17$end
#> [1] 404
#> 
#> 
#> $locations$C$items$C18
#> $locations$C$items$C18$start
#> [1] 405
#> 
#> $locations$C$items$C18$end
#> [1] 407
#> 
#> 
#> $locations$C$items$C19
#> $locations$C$items$C19$start
#> [1] 408
#> 
#> $locations$C$items$C19$end
#> [1] 409
#> 
#> 
#> $locations$C$items$C20
#> $locations$C$items$C20$start
#> [1] 410
#> 
#> $locations$C$items$C20$end
#> [1] 411
#> 
#> 
#> $locations$C$items$C21
#> $locations$C$items$C21$start
#> [1] 412
#> 
#> $locations$C$items$C21$end
#> [1] 427
#> 
#> 
#> $locations$C$items$C22
#> $locations$C$items$C22$start
#> [1] 428
#> 
#> $locations$C$items$C22$end
#> [1] 441
#> 
#> 
#> 
#> 
#> $locations$D
#> $locations$D$start
#> [1] 442
#> 
#> $locations$D$end
#> [1] 504
#> 
#> $locations$D$items
#> $locations$D$items$D1
#> $locations$D$items$D1$start
#> [1] 442
#> 
#> $locations$D$items$D1$end
#> [1] 445
#> 
#> 
#> $locations$D$items$D2
#> $locations$D$items$D2$start
#> [1] 446
#> 
#> $locations$D$items$D2$end
#> [1] 452
#> 
#> 
#> $locations$D$items$D3
#> $locations$D$items$D3$start
#> [1] 453
#> 
#> $locations$D$items$D3$end
#> [1] 454
#> 
#> 
#> $locations$D$items$D4
#> $locations$D$items$D4$start
#> [1] 455
#> 
#> $locations$D$items$D4$end
#> [1] 457
#> 
#> 
#> $locations$D$items$D5
#> $locations$D$items$D5$start
#> [1] 458
#> 
#> $locations$D$items$D5$end
#> [1] 467
#> 
#> 
#> $locations$D$items$D6
#> $locations$D$items$D6$start
#> [1] 468
#> 
#> $locations$D$items$D6$end
#> [1] 469
#> 
#> 
#> $locations$D$items$D7
#> $locations$D$items$D7$start
#> [1] 470
#> 
#> $locations$D$items$D7$end
#> [1] 471
#> 
#> 
#> $locations$D$items$D8
#> $locations$D$items$D8$start
#> [1] 472
#> 
#> $locations$D$items$D8$end
#> [1] 475
#> 
#> 
#> $locations$D$items$D9
#> $locations$D$items$D9$start
#> [1] 476
#> 
#> $locations$D$items$D9$end
#> [1] 483
#> 
#> 
#> $locations$D$items$D10
#> $locations$D$items$D10$start
#> [1] 484
#> 
#> $locations$D$items$D10$end
#> [1] 484
#> 
#> 
#> $locations$D$items$D11
#> $locations$D$items$D11$start
#> [1] 485
#> 
#> $locations$D$items$D11$end
#> [1] 490
#> 
#> 
#> $locations$D$items$D12
#> $locations$D$items$D12$start
#> [1] 491
#> 
#> $locations$D$items$D12$end
#> [1] 491
#> 
#> 
#> $locations$D$items$D13
#> $locations$D$items$D13$start
#> [1] 492
#> 
#> $locations$D$items$D13$end
#> [1] 493
#> 
#> 
#> $locations$D$items$D14
#> $locations$D$items$D14$start
#> [1] 494
#> 
#> $locations$D$items$D14$end
#> [1] 495
#> 
#> 
#> $locations$D$items$D15
#> $locations$D$items$D15$start
#> [1] 496
#> 
#> $locations$D$items$D15$end
#> [1] 496
#> 
#> 
#> $locations$D$items$D16
#> $locations$D$items$D16$start
#> [1] 497
#> 
#> $locations$D$items$D16$end
#> [1] 498
#> 
#> 
#> $locations$D$items$D17
#> $locations$D$items$D17$start
#> [1] 499
#> 
#> $locations$D$items$D17$end
#> [1] 504
#> 
#> 
#> 
#> 
#> $locations$E
#> $locations$E$start
#> [1] 505
#> 
#> $locations$E$end
#> [1] 530
#> 
#> $locations$E$items
#> $locations$E$items$E1
#> $locations$E$items$E1$start
#> [1] 505
#> 
#> $locations$E$items$E1$end
#> [1] 515
#> 
#> 
#> $locations$E$items$E2
#> $locations$E$items$E2$start
#> [1] 516
#> 
#> $locations$E$items$E2$end
#> [1] 516
#> 
#> 
#> $locations$E$items$E3
#> $locations$E$items$E3$start
#> [1] 517
#> 
#> $locations$E$items$E3$end
#> [1] 530
#> 
#> 
#> 
#> 
#> $locations$F
#> $locations$F$start
#> [1] 531
#> 
#> $locations$F$end
#> [1] 580
#> 
#> $locations$F$items
#> $locations$F$items$F1
#> $locations$F$items$F1$start
#> [1] 531
#> 
#> $locations$F$items$F1$end
#> [1] 543
#> 
#> 
#> $locations$F$items$F2
#> $locations$F$items$F2$start
#> [1] 544
#> 
#> $locations$F$items$F2$end
#> [1] 550
#> 
#> 
#> $locations$F$items$F3
#> $locations$F$items$F3$start
#> [1] 551
#> 
#> $locations$F$items$F3$end
#> [1] 560
#> 
#> 
#> $locations$F$items$F4
#> $locations$F$items$F4$start
#> [1] 561
#> 
#> $locations$F$items$F4$end
#> [1] 580
#> 
#> 
#> 
#> 
#> $locations$G
#> $locations$G$start
#> [1] 581
#> 
#> $locations$G$end
#> [1] 634
#> 
#> $locations$G$items
#> $locations$G$items$G1
#> $locations$G$items$G1$start
#> [1] 581
#> 
#> $locations$G$items$G1$end
#> [1] 608
#> 
#> 
#> $locations$G$items$G2
#> $locations$G$items$G2$start
#> [1] 609
#> 
#> $locations$G$items$G2$end
#> [1] 609
#> 
#> 
#> $locations$G$items$G3
#> $locations$G$items$G3$start
#> [1] 610
#> 
#> $locations$G$items$G3$end
#> [1] 610
#> 
#> 
#> $locations$G$items$G4
#> $locations$G$items$G4$start
#> [1] 611
#> 
#> $locations$G$items$G4$end
#> [1] 615
#> 
#> 
#> $locations$G$items$G5
#> $locations$G$items$G5$start
#> [1] 616
#> 
#> $locations$G$items$G5$end
#> [1] 623
#> 
#> 
#> $locations$G$items$G6
#> $locations$G$items$G6$start
#> [1] 624
#> 
#> $locations$G$items$G6$end
#> [1] 634
#> 
#> 
#> 
#> 
#> $locations$H
#> $locations$H$start
#> [1] 635
#> 
#> $locations$H$end
#> [1] 635
#> 
#> $locations$H$items
#> $locations$H$items$H1
#> $locations$H$items$H1$start
#> [1] 635
#> 
#> $locations$H$items$H1$end
#> [1] 682
#> 
#> 
#> $locations$H$items$H2
#> $locations$H$items$H2$start
#> [1] 683
#> 
#> $locations$H$items$H2$end
#> [1] 750
#> 
#> 
#> $locations$H$items$H4
#> $locations$H$items$H4$start
#> [1] 751
#> 
#> $locations$H$items$H4$end
#> [1] 756
#> 
#> 
#> $locations$H$items$H5
#> $locations$H$items$H5$start
#> [1] 757
#> 
#> $locations$H$items$H5$end
#> [1] 764
#> 
#> 
#> $locations$H$items$H6
#> $locations$H$items$H6$start
#> [1] 765
#> 
#> $locations$H$items$H6$end
#> [1] 775
#> 
#> 
#> $locations$H$items$H7
#> $locations$H$items$H7$start
#> [1] 776
#> 
#> $locations$H$items$H7$end
#> [1] 785
#> 
#> 
#> $locations$H$items$H8
#> $locations$H$items$H8$start
#> [1] 786
#> 
#> $locations$H$items$H8$end
#> [1] 793
#> 
#> 
#> $locations$H$items$H9
#> $locations$H$items$H9$start
#> [1] 794
#> 
#> $locations$H$items$H9$end
#> [1] 797
#> 
#> 
#> $locations$H$items$H10
#> $locations$H$items$H10$start
#> [1] 798
#> 
#> $locations$H$items$H10$end
#> [1] 800
#> 
#> 
#> $locations$H$items$H11
#> $locations$H$items$H11$start
#> [1] 801
#> 
#> $locations$H$items$H11$end
#> [1] 804
#> 
#> 
#> $locations$H$items$H12
#> $locations$H$items$H12$start
#> [1] 805
#> 
#> $locations$H$items$H12$end
#> [1] 821
#> 
#> 
#> $locations$H$items$H13
#> $locations$H$items$H13$start
#> [1] 822
#> 
#> $locations$H$items$H13$end
#> [1] 831
#> 
#> 
#> $locations$H$items$H14
#> $locations$H$items$H14$start
#> [1] 832
#> 
#> $locations$H$items$H14$end
#> [1] 839
#> 
#> 
#> $locations$H$items$H15
#> $locations$H$items$H15$start
#> [1] 840
#> 
#> $locations$H$items$H15$end
#> [1] 635
#> 
#> 
#> 
#> 
#> 
#> $values
#> $values$year
#> [1] 2008
#> 
#> $values$B1_men_full_new_undergrad
#> [1] 653
#> 
#> $values$B1_men_full_first_undergrad
#> [1] 0
#> 
#> $values$B1_men_full_other_undergrad
#> [1] 1979
#> 
#> $values$B1_men_full_degree_undergrad
#> [1] 2632
#> 
#> $values$B1_men_full_credit_undergrad
#> [1] 2
#> 
#> $values$B1_men_full_total_undergrad
#> [1] 2634
#> 
#> $values$B1_men_part_new_undergrad
#> [1] 0
#> 
#> $values$B1_men_part_first_undergrad
#> [1] 0
#> 
#> $values$B1_men_part_other_undergrad
#> [1] 3
#> 
#> $values$B1_men_part_degree_undergrad
#> [1] 3
#> 
#> $values$B1_men_part_credit_undergrad
#> [1] 7
#> 
#> $values$B1_men_part_total_undergrad
#> [1] 10
#> 
#> $values$B1_men_full_new_grad
#> [1] 755
#> 
#> $values$B1_men_full_other_grad
#> [1] 1559
#> 
#> $values$B1_men_full_degree_grad
#> [1] NA
#> 
#> $values$B1_men_full_credit_grad
#> [1] 34
#> 
#> $values$B1_men_full_total_grad
#> [1] 2348
#> 
#> $values$B1_men_part_new_grad
#> [1] 6
#> 
#> $values$B1_men_part_other_grad
#> [1] 24
#> 
#> $values$B1_men_part_degree_grad
#> [1] NA
#> 
#> $values$B1_men_part_credit_grad
#> [1] 4
#> 
#> $values$B1_men_part_total_grad
#> [1] 34
#> 
#> $values$B1_men_full_total
#> [1] NA
#> 
#> $values$B1_men_part_total
#> [1] NA
#> 
#> $values$B1_women_full_new_undergrad
#> [1] 665
#> 
#> $values$B1_women_full_first_undergrad
#> [1] 0
#> 
#> $values$B1_women_full_other_undergrad
#> [1] 1954
#> 
#> $values$B1_women_full_degree_undergrad
#> [1] 2619
#> 
#> $values$B1_women_full_credit_undergrad
#> [1] 1
#> 
#> $values$B1_women_full_total_undergrad
#> [1] 2620
#> 
#> $values$B1_women_part_new_undergrad
#> [1] 0
#> 
#> $values$B1_women_part_first_undergrad
#> [1] 0
#> 
#> $values$B1_women_part_other_undergrad
#> [1] 2
#> 
#> $values$B1_women_part_degree_undergrad
#> [1] 2
#> 
#> $values$B1_women_part_credit_undergrad
#> [1] 11
#> 
#> $values$B1_women_part_total_undergrad
#> [1] 13
#> 
#> $values$B1_women_full_new_grad
#> [1] 830
#> 
#> $values$B1_women_full_other_grad
#> [1] 1586
#> 
#> $values$B1_women_full_degree_grad
#> [1] NA
#> 
#> $values$B1_women_full_credit_grad
#> [1] 40
#> 
#> $values$B1_women_full_total_grad
#> [1] 2456
#> 
#> $values$B1_women_part_new_grad
#> [1] 9
#> 
#> $values$B1_women_part_other_grad
#> [1] 60
#> 
#> $values$B1_women_part_degree_grad
#> [1] NA
#> 
#> $values$B1_women_part_credit_grad
#> [1] 9
#> 
#> $values$B1_women_part_total_grad
#> [1] 78
#> 
#> $values$B1_women_full_total
#> [1] NA
#> 
#> $values$B1_women_part_total
#> [1] NA
#> 
#> $values$B1_other_full_new_undergrad
#> [1] NA
#> 
#> $values$B1_other_full_first_undergrad
#> [1] NA
#> 
#> $values$B1_other_full_other_undergrad
#> [1] NA
#> 
#> $values$B1_other_full_degree_undergrad
#> [1] NA
#> 
#> $values$B1_other_full_credit_undergrad
#> [1] NA
#> 
#> $values$B1_other_full_total_undergrad
#> [1] NA
#> 
#> $values$B1_other_part_new_undergrad
#> [1] NA
#> 
#> $values$B1_other_part_first_undergrad
#> [1] NA
#> 
#> $values$B1_other_part_other_undergrad
#> [1] NA
#> 
#> $values$B1_other_part_degree_undergrad
#> [1] NA
#> 
#> $values$B1_other_part_credit_undergrad
#> [1] NA
#> 
#> $values$B1_other_part_total_undergrad
#> [1] NA
#> 
#> $values$B1_other_full_new_grad
#> [1] NA
#> 
#> $values$B1_other_full_other_grad
#> [1] NA
#> 
#> $values$B1_other_full_degree_grad
#> [1] NA
#> 
#> $values$B1_other_full_credit_grad
#> [1] NA
#> 
#> $values$B1_other_full_total_grad
#> [1] NA
#> 
#> $values$B1_other_part_new_grad
#> [1] NA
#> 
#> $values$B1_other_part_other_grad
#> [1] NA
#> 
#> $values$B1_other_part_degree_grad
#> [1] NA
#> 
#> $values$B1_other_part_credit_grad
#> [1] NA
#> 
#> $values$B1_other_part_total_grad
#> [1] NA
#> 
#> $values$B1_other_full_total
#> [1] NA
#> 
#> $values$B1_other_part_total
#> [1] NA
#> 
#> $values$B1_undergrad
#> [1] 5277
#> 
#> $values$B1_grad
#> [1] 6169
#> 
#> $values$B1_total
#> [1] 11446
#> 
#> $values$G1_first_year_private
#> [1] 36500
#> 
#> $values$G1_first_year_public_district
#> [1] NA
#> 
#> $values$G1_first_year_public_state
#> [1] NA
#> 
#> $values$G1_first_year_public_out_state
#> [1] NA
#> 
#> $values$G1_first_year_non_resident
#> [1] NA
#> 
#> $values$G1_first_year_fees
#> [1] 0
#> 
#> $values$G1_first_year_food_room
#> [1] 11000
#> 
#> $values$G1_first_year_room
#> [1] NA
#> 
#> $values$G1_first_year_food
#> [1] NA
#> 
#> $values$G1_first_year_total
#> [1] NA
#> 
#> $values$G1_undergrad_private
#> [1] 36500
#> 
#> $values$G1_undergrad_public_district
#> [1] NA
#> 
#> $values$G1_undergrad_public_state
#> [1] NA
#> 
#> $values$G1_undergrad_public_out_state
#> [1] NA
#> 
#> $values$G1_undergrad_non_resident
#> [1] NA
#> 
#> $values$G1_undergrad_fees
#> [1] 0
#> 
#> $values$G1_undergrad_food_room
#> [1] 11000
#> 
#> $values$G1_undergrad_room
#> [1] NA
#> 
#> $values$G1_undergrad_food
#> [1] NA
#> 
#> $values$G1_undergrad_total
#> [1] NA
#> 
#> $values$G1_total
#> [1] NA
#> 
#> $values$G5_resident_book_supplies
#> [1] 950
#> 
#> $values$G5_resident_room
#> [1] NA
#> 
#> $values$G5_resident_food
#> [1] NA
#> 
#> $values$G5_resident_room_food
#> [1] NA
#> 
#> $values$G5_resident_transport
#> [1] 720
#> 
#> $values$G5_resident_other
#> [1] 2100
#> 
#> $values$G5_commuter_home_book_supplies
#> [1] 950
#> 
#> $values$G5_commuter_home_room
#> [1] NA
#> 
#> $values$G5_commuter_home_food
#> [1] NA
#> 
#> $values$G5_commuter_home_room_food
#> [1] NA
#> 
#> $values$G5_commuter_home_transport
#> [1] NA
#> 
#> $values$G5_commuter_home_other
#> [1] 2050
#> 
#> $values$G5_commuter_not_home_book_supplies
#> [1] 950
#> 
#> $values$G5_commuter_not_home_room
#> [1] NA
#> 
#> $values$G5_commuter_not_home_food
#> [1] NA
#> 
#> $values$G5_commuter_not_home_room_food
#> [1] NA
#> 
#> $values$G5_commuter_not_home_transport
#> [1] NA
#> 
#> $values$G5_commuter_not_home_other
#> [1] 2050
#> 
#> $values$H1_need_scholarship_federal
#> [1] 3959619
#> 
#> $values$H1_need_scholarship_state
#> [1] 274720
#> 
#> $values$H1_need_scholarship_institutional
#> [1] 87892290
#> 
#> $values$H1_need_scholarship_external
#> [1] 4522313
#> 
#> $values$H1_need_scholarship_total
#> [1] 96648942
#> 
#> $values$H1_need_self_loan
#> [1] 701076
#> 
#> $values$H1_need_self_federal
#> [1] 798400
#> 
#> $values$H1_need_self_other
#> [1] 2910835
#> 
#> $values$H1_need_self_total
#> [1] 4410311
#> 
#> $values$H1_need_parent_loan
#> [1] 0
#> 
#> $values$H1_need_waiver
#> [1] 0
#> 
#> $values$H1_need_athletic
#> [1] 0
#> 
#> $values$H1_non_need_scholarship_federal
#> [1] NA
#> 
#> $values$H1_non_need_scholarship_state
#> [1] NA
#> 
#> $values$H1_non_need_scholarship_institutional
#> [1] NA
#> 
#> $values$H1_non_need_scholarship_external
#> [1] 637383
#> 
#> $values$H1_non_need_scholarship_total
#> [1] 0
#> 
#> $values$H1_non_need_self_loan
#> [1] 1726896
#> 
#> $values$H1_non_need_self_federal
#> [1] NA
#> 
#> $values$H1_non_need_self_other
#> [1] NA
#> 
#> $values$H1_non_need_self_total
#> [1] 0
#> 
#> $values$H1_non_need_parent_loan
#> [1] 5366240
#> 
#> $values$H1_non_need_waiver
#> [1] NA
#> 
#> $values$H1_non_need_athletic
#> [1] NA
#> 
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#> 
#> $<NA>
#> NULL
#>